Refactor Kafka modules. Moves 012-kafka-streams to 301-vertx-kafka
