--- 012-quarkus-kafka-streams/src/test/java/io/quarkus/qe/AlertMonitorIT.java
@@ -1,7 +1,0 @@
-package io.quarkus.qe;
-
-import io.quarkus.test.junit.NativeImageTest;
-
-@NativeImageTest
-public class AlertMonitorIT extends AlertMonitorTest {
-}
