--- 012-quarkus-kafka-streams/src/test/java/io/quarkus/qe/StrimziKafkaResource.java
@@ -1,37 +1,0 @@
-package io.quarkus.qe;
-
-import io.quarkus.test.common.QuarkusTestResourceLifecycleManager;
-import io.strimzi.StrimziKafkaContainer;
-import org.jboss.logging.Logger;
-import org.testcontainers.containers.Network;
-
-import java.util.HashMap;
-import java.util.Map;
-
-public class StrimziKafkaResource implements QuarkusTestResourceLifecycleManager {
-
-    private static final Logger LOG = Logger.getLogger(StrimziKafkaResource.class);
-    private StrimziKafkaContainer kafkaContainer;
-
-    @Override
-    public Map<String, String> start() {
-        Network network = Network.newNetwork();
-        kafkaContainer = new StrimziKafkaContainer("0.18.0-kafka-2.5.0").withNetwork(network);
-        kafkaContainer.start();
-
-        String kafkaUrl = kafkaContainer.getBootstrapServers();
-        LOG.info(String.format("TestContainers Kafka URL -> %s", kafkaUrl));
-
-        Map<String, String> config = new HashMap<>();
-        config.put("kafka.bootstrap.servers", kafkaUrl);
-        config.put("quarkus.kafka-streams.bootstrap-servers", kafkaUrl);
-
-        return config;
-
-    }
-
-    @Override
-    public void stop() {
-        kafkaContainer.close();
-    }
-}
