+++ 301-quarkus-vertx-kafka/README.md
@@ -1,6 +1,10 @@
-# Quarkus Vertx Kafka Ext
+# Table of Contents
+1. [Quarkus Vertx Kafka Ext](#quarkus-vertx-kafka-ext)
+2. [Quarkus Kafka Streams](#quarkus-kafka-streams)
 
-## Recommended Readings
+## Quarkus Vertx Kafka Ext
+
+### Recommended Readings
 
 [Quarkus Apache Kafka Reactive Messaging](https://quarkus.io/guides/kafka)
 
@@ -12,14 +16,14 @@
 
 [Test Profiles: writing your profile](https://quarkus.io/guides/getting-started-testing#writing-a-profile)
 
-## Requirements
+### Requirements
 
 To compile and run this demo you will need:
 
 - JDK 11+
 - Docker / docker-compose
 
-## Scope of the testing
+### Scope of the testing
 
 Vert.x Kafka exploratory test. 
 * Kafka producer
@@ -38,34 +42,79 @@ so we don't really need a new verticle to consume Kafka in a reactive Non-blocki
  [1]: https://vertx.io/docs/vertx-core/java/#_verticles
  [2]: https://vertx.io/docs/vertx-core/java/#_executing_periodic_and_delayed_actions
  
-## Live coding with Quarkus
+### Live coding with Quarkus
 
-### Strimzi:
+#### Strimzi:
 
 > docker-compose -f docker-compose-strimzi.yaml up
 >
 > mvn quarkus:dev
 
-### Confluent:
+#### Confluent:
 > docker-compose -f docker-compose-confluent.yaml up
 >
 > mvn -Dquarkus.profile=confluent quarkus:dev
 
-### Run automated tests
+#### Run automated tests
 > mvn test
 
-## Troubleshooting
+### Troubleshooting
 
-### When I swap from Strimzi/confluent sometimes Kafka broker doesn't weak up (docker-compose).
+#### When I swap from Strimzi/confluent sometimes Kafka broker doesn't weak up (docker-compose).
 Is something that We should investigate, must be for some service name collisions. It's the middle time you could clean your docker containers by running the following commands:
 > docker stop $(docker ps -a -q)
 >
 > docker rm $(docker ps -a -q)
 
-### [Only Fedora] When I launch a docker compose, kafka broker can't reach Zookeper node rather a internal network issue. 
+#### [Only Fedora] When I launch a docker compose, kafka broker can't reach Zookeper node rather a internal network issue. 
 `error: kafka.zookeeper.ZooKeeperClientTimeoutException: Timed out waiting for connection while in state: CONNECTING
 
 Looks that the default backend for firewalld was changed from iptables to nftables, since Fedora32. A quick patch it's a rollback to iptables as firewalld backend.
 > sudo sed -i 's/FirewallBackend=nftables/FirewallBackend=iptables/g' /etc/firewalld/firewalld.conf
 >
 > sudo systemctl restart firewalld docker
+
+## Quarkus Kafka Streams
+
+### Recommended Readings
+
+[Quarkus Apache Kafka Streams](https://quarkus.io/guides/kafka-streams)
+
+### Requirements
+
+To compile and run this demo you will need:
+
+- JDK 11+
+- Docker / docker-compose
+
+### Scope of the testing
+
+Kafka streams 
+* Kafka producer
+* Kafka consumer
+* Kafka stream
+* Kafka stream 'windowed'
+* Automating testing (JVM/native)
+
+**Brief description:** Login monitor alerts. This kafka stream will handle login status events, and will group these events in fixed time windows. 
+So if the number of wrong access excess a threshold, then a new alert event is thrown. All aggregated events(not only unauthorized) are persisted.   
+
+![Architecture Diagram](docs/LoginMonitor.png)
+
+### Live coding with Quarkus
+
+Monitor url: `http://localhost:8080/monitor/stream`
+
+#### Strimzi:
+
+> docker-compose -f docker-compose-strimzi.yaml up
+>
+> mvn quarkus:dev
+
+#### Confluent:
+> docker-compose -f docker-compose-confluent.yaml up
+>
+> mvn quarkus:dev
+
+#### Run automated tests
+> mvn test
