+++ 301-quarkus-vertx-kafka/src/main/java/io/quarkus/qe/kafka/model/LoginAggregation.java
@@ -1,0 +1,18 @@
+package io.quarkus.qe.kafka.model;
+
+import io.quarkus.runtime.annotations.RegisterForReflection;
+
+@RegisterForReflection
+public class LoginAggregation {
+    public String endpoint;
+    public int code;
+    public int count;
+
+    public LoginAggregation updateFrom(LoginAttempt loginAttempt) {
+        endpoint = loginAttempt.endpoint;
+        code = loginAttempt.code;
+        count++;
+
+        return this;
+    }
+}
