--- 301-quarkus-vertx-kafka/src/main/java/io/quarkus/qe/vertx/kafka/config/VertxKProducerConfig.java
@@ -1,15 +1,0 @@
-package io.quarkus.qe.vertx.kafka.config;
-
-import io.quarkus.arc.config.ConfigProperties;
-import org.eclipse.microprofile.config.inject.ConfigProperty;
-
-@ConfigProperties(prefix = "vertx.kafka.producer")
-public class VertxKProducerConfig {
-
-    @ConfigProperty(name = "delaySec")
-    public long delay;
-
-    @ConfigProperty(name = "batchSize")
-    public int batchSize;
-
-}
