+++ 301-quarkus-vertx-kafka/src/main/resources/META-INF/resources/monitor.html
