+++ 301-quarkus-vertx-kafka/src/test/java/io/quarkus/qe/kafka/AlertMonitorIT.java
@@ -1,0 +1,7 @@
+package io.quarkus.qe.kafka;
+
+import io.quarkus.test.junit.NativeImageTest;
+
+@NativeImageTest
+public class AlertMonitorIT extends AlertMonitorTest {
+}
