+++ 301-quarkus-vertx-kafka/src/test/java/io/quarkus/qe/kafka/resources/ConfluentTestProfile.java
@@ -1,0 +1,20 @@
+package io.quarkus.qe.kafka.resources;
+
+import io.quarkus.qe.kafka.resources.ConfluentKafkaResource;
+import io.quarkus.test.junit.QuarkusTestProfile;
+
+import java.util.Collections;
+import java.util.List;
+
+public class ConfluentTestProfile implements QuarkusTestProfile {
+
+    @Override
+    public String getConfigProfile() {
+        return "confluent";
+    }
+
+    @Override
+    public List<TestResourceEntry> testResources() {
+        return Collections.singletonList(new TestResourceEntry(ConfluentKafkaResource.class));
+    }
+}
