--- 301-quarkus-vertx-kafka/src/test/java/io/quarkus/qe/vertx/kafka/ConfluentKafkaIT.java
@@ -1,7 +1,0 @@
-package io.quarkus.qe.vertx.kafka;
-
-import io.quarkus.test.junit.NativeImageTest;
-
-@NativeImageTest
-public class ConfluentKafkaIT extends ConfluentKafkaTest {
-}
