+++ 301-quarkus-vertx-kafka/src/test/resources/application.properties
@@ -1,4 +1,42 @@
+quarkus.test.native-image-profile=test
+
 ## Configuration file - Quarkus profile: Strimzi [Default]
+
+# Kafka Streams scenario
+
+producer.httpCodes=200,201,400,401,403,500
+producer.loginUrls=redhat/login,inditex/login,santander/login,bbva/login
+
+mp.messaging.outgoing.login-http-response-values.connector=smallrye-kafka
+mp.messaging.outgoing.login-http-response-values.key.serializer=org.apache.kafka.common.serialization.StringSerializer
+mp.messaging.outgoing.login-http-response-values.value.serializer=org.apache.kafka.common.serialization.StringSerializer
+
+mp.messaging.incoming.login-denied.connector=smallrye-kafka
+mp.messaging.incoming.login-denied.value.deserializer=org.apache.kafka.common.serialization.StringDeserializer
+mp.messaging.incoming.login-denied.key.deserializer=org.apache.kafka.common.serialization.StringDeserializer
+
+quarkus.kafka-streams.bootstrap-servers=localhost:9092
+quarkus.kafka-streams.application-id=login-denied-aggregator
+quarkus.kafka-streams.application-server=localhost:8080
+
+quarkus.kafka-streams.topics=login-http-response-values
+login.denied.windows.sec=3
+login.denied.threshold=2
+
+# streams options
+kafka-streams.cache.max.bytes.buffering=10240
+kafka-streams.commit.interval.ms=1000
+kafka-streams.metadata.max.age.ms=500
+kafka-streams.auto.offset.reset=earliest
+kafka-streams.metrics.recording.level=DEBUG
+kafka-streams.consumer.session.timeout.ms=10000
+kafka-streams.consumer.heartbeat.interval.ms=80
+
+quarkus.log.console.enable=true
+quarkus.log.console.level=INFO
+
+
+# Kafka AVRO scenario
 
 %strimzi.kafka.bootstrap.servers=localhost:9092
 %strimzi.mp.messaging.connector.smallrye-kafka.apicurio.registry.url=http://localhost:8081/api
